services/news-ingestor/requirements.txt
fastapi
uvicorn
pydantic
feedparser
requests
kafka-python
sqlite-utils
python-dateutil
